+++ Client/ParameterClient.h
@@ -9,12 +9,13 @@
  * \~english The parameters of client
  * \details
  *   - The parameters is valid in the CClient and plugin.
+ *   - The interface only is implemented by CClient and used by CClient or plugin.
  *   - Plugin accessed via CParameterConnecter::GetParameterClient .
- *   - The application cannot access it directly,
- *     it can only be set via CClient::GetSettingsWidgets .
+ *   - The application cannot access it,
+ *     but the application can be set it via CClient::GetSettingsWidgets .
  *
- * \note
- *  - The interface only is implemented by CClient and used by plugin.
+ * \subsection sub_Use_CParameterClient Use CParameterClient
+ * \subsubsection sub_Set_CParameterClient_in_CParameterConnecter Set CParameterClient for CParameterConnecter
  *  - If the parameters of connecter ( CParameterConnecter or it's derived class ) requires a CParameterClient.
  *    - Please instantiate the parameters of connecter in the constructor of the CConnecter or it's derived class.
  *    - Call CConnecter::SetParameter to set the parameters.
@@ -22,6 +23,9 @@
  *    - If you are sure to the parameter does not need CParameterClient.
  *      please overload the CConnecter::SetParameterClient in the CConnecter derived class.
  *      don't set it.
+ * \subsubsection sub_Use_CParameterClient_in_CParameterConnecter Use CParameterClient in CParameterConnecter
+ *  - Use CParameterConnecter::GetParameterClient() get CParameterClient
+ *    in the parameters of connecter ( CParameterConnecter or it's derived class )
  *  - The parameters of connecter( CParameterConnecter or it's derived class )
  *    use the value of CParameterClient as its initial value. \n
  *    For example, saving a password can be initialized with it.
@@ -29,25 +33,27 @@
  *    \see CParameterUser::onLoad
  *
  *
- * \~chinese 客户端参数
+ * \~chinese 客户端库 (CClient) 参数。仅在客户端库 (CClient) 和插件中使用。
  * \details
- * - 此参数仅在客户端和插件中使用。
+ * - 此接口仅由客户端库 (CClient) 实例化，插件和客户端库（ CClient ) 使用。
  * - 插件通过 CParameterConnecter::GetParameterClient 访问。
- * - 应用程序不能直接访问，只能通过 CClient::GetSettingsWidgets 进行设置。
+ * - 应用程序不能访问，但是可以通过 CClient::GetSettingsWidgets 进行设置。
  *
- * \note
- *  - 此接口仅由 CClient 实现，插件和客户端（ CClient ) 使用。
- *  - 如果连接参数 ( CParameterConnecter 或其派生类 ) 需要 CParameterClient 。
- *    - 请在 CConnecter 派生类的构造函数中实例化连接参数，
+ * \subsection sub_Use_CParameterClient 使用客户端库参数 (CParameterClient)
+ * \subsubsection sub_Set_CParameterClient_in_CParameterConnecter 为连接参数 ( CParameterConnecter 或其派生类 ) 设置客户端库参数 (CParameterClient)
+ *  - 如果连接参数 ( CParameterConnecter 或其派生类 ) 需要客户端库参数 (CParameterClient) 。
+ *    - 请在 CConnecter 派生类的构造函数中实例化连接参数。
  *    - 调用 CConnecter::SetParameter 设置参数指针。
  *    - 默认会自动为连接参数设置 CParameterClient 。详见: CClient::CreateConnecter 。
  *    - 如果参数不需要 CParameterClient ，
  *      那请在 CConnecter 派生类重载 CConnecter::SetParameterClient 不设置它。
- *  - 连接参数 ( CParameterConnecter 或其派生类 ) 使用 CParameterClient 的值做为其初始值。
- *    请在 CParameterConnecter::onLoad 中初始化。\n
- *    例如：保存密码以 CParameterClient 的成员值做为其初始化值。
- *    \snippet Client/ParameterCompone/ParameterUser.cpp Initialize parameter
- *    \see CParameterUser::onLoad
+ * \subsubsection sub_Use_CParameterClient_in_CParameterConnecter 在连接参数 ( CParameterConnecter 或其派生类 ) 中使用客户端库参数 (CParameterClient)
+ *   - 在连接参数 ( CParameterConnecter 或其派生类 ) 使用 CParameterConnecter::GetParameterClient() 来得到客户端库参数 (CParameterClient) 。
+ *   - 连接参数 ( CParameterConnecter 或其派生类 ) 使用客户端库参数 (CParameterClient) 的值做为其初始值。
+ *     请在 CParameterConnecter::onLoad 中初始化。\n
+ *     例如：保存密码以 CParameterClient 的成员值做为其初始化值。
+ *     \snippet Client/ParameterCompone/ParameterUser.cpp Initialize parameter
+ *     \see CParameterUser::onLoad
  *
  * \~
  * \see CClient::CreateConnecter