+++ ChangeLog.md
@@ -6,6 +6,7 @@
   - Modify debian package
   - Use [RabbitCommon v2.2.0](https://github.com/KangLin/RabbitCommon/releases/tag/v2.2.0)
   - Add SSH tunnel
+  - Refactor parameter
 
 ### v0.0.26
 - Use RabbitCommon v2